requirements.txt
fastapi[standard]
datacsv
python-dotenv
boto3